+++ src/lifecycle/index.ts
@@ -1,5 +1,6 @@
 import * as checkUpdate from './check-update';
 import * as common from './common';
+import * as contractMigrations from './contract-migrations';
 import * as rollback from './rollback';
 import * as uninstall from './uninstall';
 import * as update from './update';
@@ -9,6 +10,7 @@ import * as updateTrust from './update-trust';
 export {
     checkUpdate,
     common,
+    contractMigrations,
     rollback,
     uninstall,
     update,